Refactor MarseyPatcher.cs, split it into multiple classes.
