+++ SS14.Loader/SS14.Loader.csproj
@@ -18,7 +18,7 @@
     <Compile Include="../SS14.Launcher/Utility/ZStd.cs" />
     <Compile Include="../SS14.Launcher/Utility/Libc.cs" />
     <Compile Include="../SS14.Launcher/Utility/SqliteBlobStream.cs" />
-    <Compile Include="../SS14.Launcher/Marsey/MarseyPatcher.cs" />
+    <Compile Include="../SS14.Launcher/Marsey/*" />
   </ItemGroup>
 
   <ItemGroup>
